git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@4959 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/Utilities/CodeUtility.h
@@ -14,7 +14,7 @@ static bool StringToBool( const std::string& value);
 
 static inline long nint( double value )
 { 
-   return floor( 0.5 + value);
+   return (long)floor( 0.5 + value);
 };
